Dropped patches.kabi/
New backport of patches.suse/block-Do-not-reread-partition-table-on-exclusively-o.patch

Conflicts:
	blacklist.conf: dropped kABI items
	patches.suse/drm-amdkfd-Check-for-null-pointer-after-calling-kmem.patch: merged References:
	patches.suse/net-mlx5-Fix-mlx5_get_next_dev-peer-device-matching.patch: --ours
	patches.suse/net-mlx5-Lag-filter-non-compatible-devices.patch: --ours
	patches.suse/timers-implement-usleep_idle_range.patch: dropped duplicate
	series.conf: git-sort

suse-commit: 4be84f8e6e4f81f2b78d9f6438fa7c9394ab9439

+++ kernel/trace/trace_osnoise.c
@@ -731,7 +731,7 @@ void osnoise_trace_irq_entry(int id)
 void osnoise_trace_irq_exit(int id, const char *desc)
 {
 	struct osnoise_variables *osn_var = this_cpu_osn_var();
-	int duration;
+	s64 duration;
 
 	if (!osn_var->sampling)
 		return;
@@ -862,7 +862,7 @@ static void trace_softirq_entry_callback(void *data, unsigned int vec_nr)
 static void trace_softirq_exit_callback(void *data, unsigned int vec_nr)
 {
 	struct osnoise_variables *osn_var = this_cpu_osn_var();
-	int duration;
+	s64 duration;
 
 	if (!osn_var->sampling)
 		return;
@@ -970,7 +970,7 @@ thread_entry(struct osnoise_variables *osn_var, struct task_struct *t)
 static void
 thread_exit(struct osnoise_variables *osn_var, struct task_struct *t)
 {
-	int duration;
+	s64 duration;
 
 	if (!osn_var->sampling)
 		return;
